requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
decorator==5.1.0
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
drf-yasg==1.20.0
flex==6.14.1
idna==3.3
inflection==0.5.1
ipython==7.29.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
jsonpointer==2.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
packaging==21.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.21
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
pymongo==3.12.1
pyparsing==2.4.7
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rfc3987==1.3.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.2.4
strict-rfc3339==0.7
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
validate-email==1.3
wcwidth==0.2.5
